move cans.py to models package
